+++ 02/ex02/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: dogwak <dogwak@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/25 18:08:46 by dogwak            #+#    #+#             */
-/*   Updated: 2024/09/26 01:04:52 by dogwak           ###   ########.fr       */
+/*   Updated: 2024/10/17 20:02:44 by dogwak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -171,7 +171,7 @@ float Fixed::toFloat() const
 	return static_cast<float>(bitData) / static_cast<float>(1 << binaryPoint);
 }
 
-float Fixed::toInt() const
+int Fixed::toInt() const
 {
 	return (bitData >> binaryPoint);
 }
